requirements.txt
aiohttp==3.6.1
aiopg==0.16.0
alchimia==0.8.1
alembic==1.2.0
aniso8601==8.0.0
apispec==1.3.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.0
cached-property==1.5.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.5
django-postgrespool==0.3.1
djangorestframework==3.10.3
Faker==0.9.1
Flask==1.1.1
Flask-AppBuilder==2.1.13
Flask-Babel==0.12.2
Flask-JWT-Extended==3.23.0
Flask-Login==0.4.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-OpenID==1.2.5
Flask-RESTful==0.3.7
flask-restplus==0.13.0
Flask-Script==2.0.6
flask-sqlacodegen==1.1.6.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
greenlet==0.4.15
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
inflect==2.1.0
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.19.0
meinheld==1.0.1
mixer==6.1.3
mock==3.0.5
more-itertools==7.2.0
muffin==0.32.8
muffin-jinja2==0.2.0
muffin-peewee==1.3.0
multidict==4.5.2
packaging==19.2
parso==0.5.1
peewee==2.10.2
peewee-async==0.5.12
peewee-migrate==0.14.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prison==0.1.2
prompt-toolkit==2.0.9
psutil==5.6.3
psycopg2==2.8.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.3
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
sqlacodegen==2.1.0
SQLAlchemy==1.3.8
SQLAlchemy-Utils==0.34.2
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
treq==18.6.0
Twisted==19.7.0
ujson==1.35
urllib3==1.25.5
waitress==1.3.1
wcwidth==0.1.7
WebOb==1.8.5
WebTest==2.0.33
Werkzeug==0.16.0
wheezy.caching==0.1.114
wheezy.core==0.1.140
wheezy.html==0.1.147
wheezy.http==0.1.344
wheezy.routing==0.1.157
wheezy.security==0.1.64
wheezy.validation==0.1.135
wheezy.web==0.1.485
WTForms==2.2.1
yarl==1.3.0
zipp==0.6.0
zope.interface==4.6.0
